// apps/dashboard/src/enums/Routing.ts
export enum Routing {
  Dashboard = '/dashboard',
  Workspaces = '/workspaces',
  Keys = '/keys',
  Images = '/images',
  Registries = '/registries',
  Organizations = '/members',
  UserOrganizationInvitations = '/user/invitations',
  Usage = '/usage',
  Billing = '/billing',
  LandingPage = '/',
  Logout = '/logout',
  DocsRedirect = '/docs',
  Sandboxes = '/sandboxes',
  Settings = '/settings',
  DashboardBilling = '/dashboard/billing',
  DaytonaDocs = 'https://daytona.io/docs',
  DaytonaSlack = 'https://go.daytona.io/slack',
}